pyproject.toml
[project]
name = "Agentic-AI-Financial-Agent"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Abinav Ramesh", email = "abinavrameshs@gmail.com"},
]
dependencies = ["python-dotenv>=1.0.1", "yfinance>=0.2.53", "packaging>=24.2", "duckduckgo-search>=7.4.2", "fastapi>=0.115.8", "uvicorn>=0.34.0", "groq>=0.18.0", "openai>=1.63.0", "python-multipart>=0.0.20", "crawl4ai>=0.4.247", "agno>=1.1.3", "google-genai>=1.2.0"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "GNU"}

[project.optional-dependencies]
youtube = [
    "youtube-transcript-api>=0.6.2",
]
web_scraping = [
    "crawl4ai>=0.4.247",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
